LayoutTests/imported/w3c/web-platform-tests/cors/status-preflight-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?1&code=200&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?2&code=204&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?3&code=400&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?4&code=401&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?5&code=200&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?6&code=204&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?7&code=400&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?8&code=401&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?9&code=501&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?10&code=699&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?11&code=204&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?12&code=400&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?13&code=401&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?14&code=404&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?15&code=699&headers=x-nonsimple&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/status.py?16&code=501&headers=x-nonsimple&preflight=200 due to access control checks.
Status after preflight


FAIL CORS - status after preflight on GET 200 assert_equals: response status expected 200 but got 0
FAIL CORS - status after preflight on GET 204 assert_equals: response status expected 204 but got 0
FAIL CORS - status after preflight on GET 400 assert_equals: response status expected 400 but got 0
FAIL CORS - status after preflight on GET 401 assert_equals: response status expected 401 but got 0
FAIL CORS - status after preflight on HEAD 200 assert_equals: response status expected 200 but got 0
FAIL CORS - status after preflight on HEAD 204 assert_equals: response status expected 204 but got 0
FAIL CORS - status after preflight on HEAD 400 assert_equals: response status expected 400 but got 0
FAIL CORS - status after preflight on HEAD 401 assert_equals: response status expected 401 but got 0
FAIL CORS - status after preflight on HEAD 501 assert_equals: response status expected 501 but got 0
FAIL CORS - status after preflight on HEAD 699 assert_equals: response status expected 699 but got 0
FAIL CORS - status after preflight on POST 204 assert_equals: response status expected 204 but got 0
FAIL CORS - status after preflight on POST 400 assert_equals: response status expected 400 but got 0
FAIL CORS - status after preflight on POST 401 assert_equals: response status expected 401 but got 0
FAIL CORS - status after preflight on POST 404 assert_equals: response status expected 404 but got 0
FAIL CORS - status after preflight on PUT 699 assert_equals: response status expected 699 but got 0
FAIL CORS - status after preflight on CHICKEN 501 assert_equals: response status expected 501 but got 0
